requirements.txt
numpy>=1.22.0
imageio>=2.19.0
trimesh>=3.9.0
matplotlib>=3.5.0
Pillow>=9.0.0
